chore:alert events page refresh page data every five minutes; add "skipped" Alert Validity (#242)
Improves the Alert Events page by adding a refresh interval, updating the display of alert validity, and adjusting the modal width.

Chores:

Refreshes the alert events data every five minutes to ensure the displayed information is up-to-date.
Updates the Alert Validity display to include a 'skipped' status in addition to the existing 'unknown' status.
Adjusts the width of the workflow modal to 80% of the viewport width.